[Bronze V] Title: 이상한 기호, Time: 64 ms, Memory: 5500 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cs
@@ -5,12 +5,12 @@ public class Program
 
     static void Main(string[] args)
     {
-        int numA;
-        int numB;
-        int result;
-        int[] input;
+        long numA;
+        long numB;
+        long result;
+        long[] input;
 
-        input = Array.ConvertAll(SR.ReadLine().Split(), int.Parse);
+        input = Array.ConvertAll(SR.ReadLine().Split(), long.Parse);
         numA = input[0];
         numB = input[1];
